+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Added commenter support
 - Added formatting and code style settings
 - Added uniform hint documentation support
+- Added inlay parameter hints feature
 
 ## 0.6.0-beta (2026-02-07)
 - Added parameter info support